+++ examples/image.rs
@@ -2,9 +2,10 @@
 extern crate quicksilver;
 
 use quicksilver::{
-    run, Asset, Result, State,
+    Result,
     geom::Shape,
-    graphics::{Background::Img, Color, Image, Window, WindowBuilder}
+    graphics::{Background::Img, Color, Image, Window, WindowBuilder},
+    lifecycle::{Asset, State, run},
 };
 
 struct ImageViewer {